% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEVOLUTION.R
\name{load_matrix}
\alias{load_matrix}
\title{Load a Matrix}
\usage{
load_matrix(filename, sheetname)
}
\arguments{
\item{filename}{The name of the excel document.}

\item{sheetname}{The sheetname.}
}
\value{
A matrix
}
\description{
This function loads a file as a matrix.
}
