% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEVOLUTION.R
\name{ml_tree}
\alias{ml_tree}
\title{Maximum Likelihood}
\usage{
ml_tree(Eventmatrix, root)
}
\arguments{
\item{Eventmatrix}{The phyDat-transformed EM.}

\item{root}{The root we want for the phylogeny.}
}
\value{
The ML tree object.
}
\description{
Creating the ML object.
}
