% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEVOLUTION.R
\name{mp_tree}
\alias{mp_tree}
\title{Maximum parsimony}
\usage{
mp_tree(Eventmatrix, root)
}
\arguments{
\item{Eventmatrix}{The phyDat-transformed EM.}

\item{root}{The root we want for the phylogeny.}
}
\value{
The MP tree object.
}
\description{
#' Creating the MP object.
}
