% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEVOLUTION.R
\name{MP_treeplot}
\alias{MP_treeplot}
\title{Visualising the MP-tree.}
\usage{
MP_treeplot(MP_tree, limitmp, col)
}
\arguments{
\item{MP_tree}{The MP-tree object  obtained from mp_tree()}

\item{limitmp}{The size of the plotting environment. Choose so that the entire phylogeny is clearly visible.}

\item{col}{If the phylogeny should be colored or not.}
}
\value{
The MP tree.
}
\description{
#' Visualizing the MP-tree object.
}
