% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEVOLUTION.R
\name{simplify}
\alias{simplify}
\title{Adding a stem to the data.}
\usage{
simplify(datasegment)
}
\arguments{
\item{datasegment}{The segment file.}
}
\value{
A truncated segment file.
}
\description{
Manually adding a stem to the matrix. Often not needed.
}
