% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEVOLUTION.R
\name{simplify_matrix}
\alias{simplify_matrix}
\title{Simplifying the input dataset. Useful for WGS and WES data.}
\usage{
simplify_matrix(datasegment)
}
\arguments{
\item{datasegment}{The segment file.}
}
\value{
A truncated segment file.
}
\description{
Simplifying the input dataset. Useful for WGS and WES data.
}
