% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEVOLUTION.R
\name{splitdata}
\alias{splitdata}
\title{Splitting the input file.}
\usage{
splitdata(file, name)
}
\arguments{
\item{file}{The matrix with the genetic data i.e. the output of load_matrix.}

\item{name}{The name of the tumor(s) that should be used for further analysis.}
}
\value{
The processed matrix.
}
\description{
Splitting the input excel file based on which tumors to analyze.
}
