% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEVOLUTION.R
\name{stem}
\alias{stem}
\title{Adding a stem to the data.}
\usage{
stem(eventmatrix, co, root)
}
\arguments{
\item{eventmatrix}{The output from the DEVOLUTION function.}

\item{co}{The cutoff. The proportion of cells that have to have a particular abberation for it to be considered a stem event.}

\item{root}{If a particular subclone should be the root. Choose between "Stem", "Normal" or "None".}
}
\value{
The EM together with a stem.
}
\description{
Manually adding a stem to the matrix. Often not needed.
}
