% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEVOLUTION.R
\name{subclones}
\alias{subclones}
\title{Making new subclones.}
\usage{
subclones(
  EM_test,
  file_samples_subclones,
  root,
  possible_mothers,
  cutoff,
  names,
  simplified
)
}
\description{
Making new subclones.
}
